refactor(gateway): move ApiContext to gateway/types.ts — break cycle 2
gateway/api.ts → gateway/files.ts の循環依存を解消。
ApiContext を gateway/types.ts に分離し、api.ts は re-export、
files.ts は types.ts から直接 import するよう変更。

TASK-019(ES-011)
