+++ packages/jimmy/src/gateway/types.ts
@@ -1,0 +1,12 @@
+import type { SessionManager } from "../sessions/manager.js";
+import type { Connector, JinnConfig } from "../shared/types.js";
+
+export interface ApiContext {
+  config: JinnConfig;
+  sessionManager: SessionManager;
+  startTime: number;
+  getConfig: () => JinnConfig;
+  emit: (event: string, payload: unknown) => void;
+  connectors: Map<string, Connector>;
+  reloadConnectorInstances?: () => Promise<{ started: string[]; stopped: string[]; errors: string[] }>;
+}
